Subclassed Variable
